fix multiple charts not showing
