+++ example/web/index.html
@@ -19,7 +19,7 @@
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.24.0/moment-with-locales.min.js"></script>
   <script src="https://cdn.jsdelivr.net/gh/vramosx/flutter_web_chartjs@master/js/chart.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/gh/vramosx/flutter_web_chartjs@master/js/chartjs.wrapper.utils.min.js" type="application/javascript"></script>
+  <script src="http://localhost:5000/chartjs.wrapper.utils.js" type="application/javascript"></script>
 </head>
 <body>
   <!-- This script installs service_worker.js to provide PWA functionality to
